Better handling of input actions
My original idea of every module handling its own actions and
subscriptions was probably a bad idea. It would lead to a lot
of repetetion on the main game update function and also make
hard to know what actions the player can do.

The new design centralizes the handler of actions on a separate model,
exposing only Player Actions to the outside.